Create create-strap-app pacakge
