+++ packages/create-strapi-app/README.md
@@ -1,0 +1,1 @@
+# Create strapi app
