+++ packages/strapi-generate-new/lib/utils/stop-process.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function stopProcess(message) {
+  console.error(message);
+  process.exit(1);
+};
