+++ packages/strapi-hook-mongoose/package.json
@@ -19,7 +19,6 @@
     "mongoose": "^5.5.9",
     "mongoose-float": "^1.0.4",
     "pluralize": "^7.0.0",
-    "rimraf": "^2.6.3",
     "strapi-utils": "3.0.0-beta.6"
   },
   "author": {
